feat(rankings): move strategies to component
