--- src/utils/rankingStrategies.js
@@ -1,27 +1,0 @@
-import _ from 'lodash';
-
-function baseStrategy(communes, sortKeyPath, n = 10, reverse = true) {
-  let selectedCommunes = _.sortBy(communes, commune => _.get(commune, sortKeyPath));
-  if (reverse) {
-    _.reverse(selectedCommunes);
-  }
-  selectedCommunes = _.take(selectedCommunes, n);
-
-  return selectedCommunes;
-}
-
-export function mostAbsoluteActiveCases(communes) {
-  return baseStrategy(communes, 'activos.value');
-}
-
-export function mostRelativeActiveCases(communes) {
-  return baseStrategy(communes, 'tasaActivos.value');
-}
-
-export function deltaAbsoluteActiveCases(communes) {
-  return baseStrategy(communes, 'delta.activos.value');
-}
-
-export function deltaRelativeActiveCases(communes) {
-  return baseStrategy(communes, 'delta.tasaActivos.value');
-}
